// SharedLab-EquipmentBookingSystem_Frontend-DEV/SharedLab-EquipmentBookingSystem_Frontend-DEV/src/app/core/api/system.models.ts
export interface LabRoomResponse {
  labId: number
  labName: string
  roomCode: string
  location: string
  capacity: number
  status: string
  imageUrl?: string | null
}

export interface LabRoomDetailResponse extends LabRoomResponse {
  description: string | null
  imageUrl: string | null
  usageGuideline: string | null
  managerName: string | null
}

export interface EquipmentResponse {
  equipmentId: number
  labId: number
  equipmentName: string
  status: string
  imageUrl?: string | null
}